Mirror test layout to src layout
